requirements.txt
Django==1.11
djangorestframework==3.6.2
psycopg2==2.7.1
pytz==2017.2
